Use 'git commit-tree' to sync master to release
Using 'git merge' for this really isn't appropriate, as it introduces the
possibility of merge conflicts, when what we really want is "make a new commit
on this branch with exactly the content in the master branch". We update the
Docker image tags right after that operation, and the back-and-forth updates
from tagged images to 'latest' to the next tag isn't a problem.

Use 'git commit-tree' to do what's mentioned above: make a commit with parents
of the tip of the 'release' branch and the 'master' version which will be
tagged, then update Docker image tags as before.

If there's no 'release' branch when the script is executed (e.g. for the first
release of a pipeline), we'll end up running 'git commit-tree' with the same
commit specified as both parents -- luckily Git does the right thing in this
situation and only includes the parent once. (At least, as per the output of
'git cat-file -p'.)